CSV SRS reprojection, also case mismatches in CSV headers.
CSV files are now reprojected when extracted, to deal with a
few unusual sources that have the columns in non-WGS84.

The conform spec may name lat/lon with the wrong case, so we
do our best to find it anyway. Also make sure that sources with
columns named X and Y don't cause problems.