Shapefile reading is now encoding-aware. Fixes issue #42.
This change got complicated, read issue #42 for all the detail.
The strategy here is to hope OGR created UTF-8 byte strings for
us and convert them to Unicode. If it wasn't able to do that we
rely on the user encoding.

There are no new tests checked in because I haven't found a
good small sample test case. lake-man-utf8 was updated to
explicitly supply an encoding. That file is a case where OGR
can't guess and our new behavior is to assume iso-8859-1 in
that case.

The following non-ASCII sources were hand verified to work:
be-flanders be-wa-brussels-fr be-wa-brussels-nl
ca-qc-gatineau pl-lubuskie

This code explicitly does not support Python 3. It shouldn't be
hard to make it work in Python 3; usually OGR will provide a
correct Unicode string, so we can just skip decoding it.
If OGR can't guess the encoding correctly and it's not UTF-8,
OGR currently throws an exception in Python 3.